+++ src/pserver/global_anomaly_stats.cpp
@@ -1,4 +1,4 @@
-#include <chimbuko/global_anomaly_stats.hpp>
+#include <chimbuko/pserver/global_anomaly_stats.hpp>
 
 using namespace chimbuko;
