+++ test/test_param.cpp
@@ -1,5 +1,5 @@
 #include "chimbuko/param/sstd_param.hpp"
-#include "chimbuko/global_anomaly_stats.hpp"
+#include "chimbuko/pserver/global_anomaly_stats.hpp"
 #include "chimbuko/message.hpp"
 #include <gtest/gtest.h>
 #include <random>
